[Modify] 회원가입 스텝 수정
- 휴대폰 번호 입력 단계 제거
