+++ src/pages/Register/index.ts
@@ -1,4 +1,3 @@
 export { default as Step1 } from '@/pages/Register/Step1';
 export { default as Step2 } from '@/pages/Register/Step2';
 export { default as Step3 } from '@/pages/Register/Step3';
-export { default as Step4 } from '@/pages/Register/Step4';
